detail css 수정 &  파일구조변경
