+++ css/detail.css
@@ -10,20 +10,21 @@ body {
 }
 
 header {
-    padding: 20px;
+
+    background-color: #ccc;
+    padding: 30px 0;
 }
 
-header ul li:first-child a {
-    padding: 10px;
-    background-color: #ccc;
-    color: #000;
-    border-radius: 5px;
-    transition: 0.2s;
+header h1 {
+    text-align: center;
+    font-weight: bold;
+    font-size: 24px;
 }
 
-header ul li:first-child a:hover {
-    background-color: #ccc;
-    color: #000;
+header a {
+    text-decoration: none;
+    color: #333;
+
 }
 
 .content {
@@ -142,25 +143,46 @@ header ul li:first-child a:hover {
     display: flex;
     flex-direction: column;
     gap: 20px;
-    max-width: 1080px;  
+    max-width: 1080px;
     width: 100%;
     margin-bottom: 60px;
 }
 
 .review {
-  width: calc(100% -32px);
-   background-color: #F6EAC2;
-   border-radius: 5px;
-   padding: 16px;
-   margin: 0 20px;
+    width: calc(100% -32px);
+    background-color: #F6EAC2;
+    border-radius: 5px;
+    padding: 16px;
+    margin: 0 20px;
 }
 
 .review button {
     cursor: pointer;
+    background-color: #e9daf4;
+    border: none;
+    padding: 6px 12px;
+    border-radius: 5px;
+    margin-right: 8px;
 }
-.review .comment-contnets {
+
+.review .comment-contnets, .review .comment-writer {
     margin-bottom: 12px;
 }
+
+.review input {
+    outline: none;
+    height: 30px;
+    width: 400px;
+    margin-bottom: 12px;
+    display: block;
+}
+.review textarea {
+    margin-bottom: 12px;
+    height: 50px;
+    width: 400px;
+    display: block;
+}
+
 @media (max-width: 768px) {
 
     #movie-card>div {
@@ -206,19 +228,23 @@ header ul li:first-child a:hover {
         margin: 0;
         width: 250px;
     }
+
     #contents {
         outline: none;
         width: 250px;
     }
+
     #comment-btn {
         font-size: 16px;
     }
 }
+
 @media (max-width: 500px) {
     .input-wrap input {
         margin: 0;
         width: 200px;
     }
+
     #contents {
         outline: none;
         width: 200px;
