+++ detail.html
@@ -13,11 +13,9 @@
     <script defer src="js/detail.js" type="module" ></script>
 </head>
 <body>
-    <header>
-        <ul>
-            <li><a href="/">메인으로 돌아가기</a></li>
-        </ul>
-    </header>
+  <header>
+    <h1><a href="/">NBC THEATER</a></h1>
+  </header>
     <main>
         <section class="content">
           <div id="movie-card">
@@ -29,7 +27,7 @@
             <div class="input-wrap">
               <div>
                 <label for="name">작성자</label>
-                <input id="name" type="text" placeholder="이름을 입력하세요."/>
+                <input autocomplete="off" id="name" type="text" placeholder="이름을 입력하세요."/>
               </div>
               <div>
                 <label for="password">비밀번호</label>
