+++ index.html
@@ -15,18 +15,21 @@
       
 </head>
 <body>
-
+    <header>
+      <h1><a href="/">NBC THEATER</a></h1>
+    </header>
     <main>
-        
         <section class="content">
+          <div class="map-wrap">
+            <h3>내 근처 영화관</h3>
+            <div id="map" style="width:100%;height:350px;"></div>
+          </div>
           <div class="search-container">
             <label for="search-input">영화 검색기</label>
             <input id="search-input" type="text" placeholder="영화제목을 입력하세요!"/>
             <button id="search-btn">검색</button>
           </div>
-          <div class="map-wrap">
-            <div id="map" style="width:100%;height:350px;"></div>
-          </div>
+         
           <ul id="now-play">
               <!-- 데이터삽입 -->
           </ul>
