+++ js/comment/updateComment.js
@@ -1,8 +1,8 @@
-import getComment from "./getComment.js";
-import { getData } from "../getData.js";
+import commentTemplate from "../template/commentTemplate.js";
+import { getComment } from "./getComment.js";
 
 export default function update({ e, movieId }) {
-  const data = getData(movieId);
+  const data = getComment(movieId);
   const index = e.target.parentNode.parentNode.getAttribute("data-number");
   const passwordPrompt = prompt("비밀번호를 입력해주세요");
   if (passwordPrompt !== data[index].password) {
@@ -10,9 +10,11 @@ export default function update({ e, movieId }) {
   }
   const li = e.target.parentNode.parentNode;
   li.innerHTML = `
+        <label>작성자</label>
         <input class="update-writer${index}" type="text" />
+        <label>내용</label>
         <textarea class="update-contents${index}"  placeholder="리뷰내용을 입력하세요."></textarea>
-        <button class="update-submit-btn${index}">수정제출하기</button>
+        <button class="update-submit-btn${index}">수정하기</button>
       `;
   const updateSubmitBtn = document.querySelector(`.update-submit-btn${index}`);
   const updateWriter = document.querySelector(`.update-writer${index}`);
@@ -21,9 +23,6 @@ export default function update({ e, movieId }) {
   updateContents.value = data[index].contents;
   updateWriter.focus();
   updateSubmitBtn.addEventListener("click", function () {
-    console.log(updateWriter.value);
-    console.log(updateContents.value);
-    console.log(data);
     data[index] = {
       ...data[index],
       writer: updateWriter.value,
@@ -31,6 +30,6 @@ export default function update({ e, movieId }) {
     };
     localStorage.setItem(`comment${movieId}`, JSON.stringify(data));
 
-    getComment(data);
+    commentTemplate(data);
   });
 }
