--- js/getData.js
@@ -1,4 +1,0 @@
-export const getData = function (movieId) {
-  const reviewData = JSON.parse(localStorage.getItem(`comment${movieId}`));
-  return reviewData;
-};
